fix: prevent automatic image analysis for messages without command prefix #8
